chore(ui): move form util types to separate interface file
