+++ openmetadata-ui/src/main/resources/ui/src/components/AddService/Steps/ConfigureService.tsx
@@ -14,8 +14,9 @@
 import { Button, Form, FormProps, Space } from 'antd';
 import { ENTITY_NAME_REGEX } from 'constants/regex.constants';
 import { t } from 'i18next';
+import { FieldProp, FieldTypes } from 'interface/genericForm.interface';
 import React from 'react';
-import { FieldProp, FieldTypes, generateFormFields } from 'utils/formUtils';
+import { generateFormFields } from 'utils/formUtils';
 import { ConfigureServiceProps } from './Steps.interface';
 
 const ConfigureService = ({
