+++ openmetadata-ui/src/main/resources/ui/src/components/common/DBTConfigFormBuilder/DBTCloudConfig.tsx
@@ -12,8 +12,9 @@
  */
 
 import { t } from 'i18next';
+import { FieldProp, FieldTypes } from 'interface/genericForm.interface';
 import React, { Fragment, FunctionComponent } from 'react';
-import { FieldProp, FieldTypes, generateFormFields } from 'utils/formUtils';
+import { generateFormFields } from 'utils/formUtils';
 import DBTCommonFields from './DBTCommonFields.component';
 import { DbtConfigCloud } from './DBTConfigForm.interface';
