+++ openmetadata-ui/src/main/resources/ui/src/components/common/DBTConfigFormBuilder/DBTGCSConfig.tsx
@@ -12,13 +12,9 @@
  */
 
 import { t } from 'i18next';
+import { FieldProp, FieldTypes } from 'interface/genericForm.interface';
 import React, { Fragment, FunctionComponent } from 'react';
-import {
-  FieldProp,
-  FieldTypes,
-  generateFormFields,
-  getField,
-} from 'utils/formUtils';
+import { generateFormFields, getField } from 'utils/formUtils';
 import { GCSCredentialsValues } from '../../../generated/metadataIngestion/dbtPipeline';
 import DBTCommonFields from './DBTCommonFields.component';
 import { DbtConfigS3GCS } from './DBTConfigForm.interface';
