+++ openmetadata-ui/src/main/resources/ui/src/pages/EditCustomLogoConfig/EditCustomLogoConfig.tsx
@@ -28,6 +28,7 @@ import {
 import { ServiceCategory } from 'enums/service.enum';
 import { LogoConfiguration } from 'generated/configuration/applicationConfiguration';
 import { Settings, SettingType } from 'generated/settings/settings';
+import { FieldProp, FieldTypes } from 'interface/genericForm.interface';
 import React, { useEffect, useMemo, useState } from 'react';
 import { useTranslation } from 'react-i18next';
 import { useHistory } from 'react-router-dom';
@@ -35,7 +36,7 @@ import {
   getSettingsConfigFromConfigType,
   updateSettingsConfig,
 } from 'rest/settingConfigAPI';
-import { FieldProp, FieldTypes, generateFormFields } from 'utils/formUtils';
+import { generateFormFields } from 'utils/formUtils';
 import { getSettingPath } from 'utils/RouterUtils';
 import { showErrorToast, showSuccessToast } from 'utils/ToastUtils';
