+++ openmetadata-ui/src/main/resources/ui/src/pages/TagsPage/TagsForm.tsx
@@ -15,9 +15,10 @@ import { Form, Modal, Typography } from 'antd';
 import { VALIDATION_MESSAGES } from 'constants/constants';
 import { ENTITY_NAME_REGEX } from 'constants/regex.constants';
 import { DEFAULT_FORM_VALUE } from 'constants/Tags.constant';
+import { FieldProp, FieldTypes } from 'interface/genericForm.interface';
 import React, { useEffect } from 'react';
 import { useTranslation } from 'react-i18next';
-import { FieldProp, FieldTypes, generateFormFields } from 'utils/formUtils';
+import { generateFormFields } from 'utils/formUtils';
 import { RenameFormProps } from './TagsPage.interface';
 
 const TagsForm = ({
